Added Async to Trim Panel Component!!
Added async to the ankylosaurus project. Implemented it on the trim panel component, however this has a few kinks to workout. The trim component needs an improved algorithm for testing the trims for starters.

And the way async computes the time progress is completely wrong, so will need to figure that out